Changed ordering method for configs
When the card ID is used as the configs key, calling Object.keys() on
it means it’s iterated in a different order to the user’s hand. Using
the card index in configs instead presents a different problem - when a
new card is added, the previous object is simply overwritten. Instead,
I’m keeping the hand in order by using both the idx and the card ID as
the key.